+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
 	"cSpell.words": [
+		"muutils",
 		"zanj"
 	]
 }
