+++ muutils/json_serialize/__init__.py
@@ -1,0 +1,1 @@
+from muutils.json_serialize.json_serialize import json_serialize, JsonSerializer, JSONitem
