+++ muutils/zanj/__init__.py
@@ -1,0 +1,1 @@
+from muutils.zanj.zanj import ZANJ
